Sorting of the functions of graphs.py into analysis and plotting
