+++ hcga/graphs.py
@@ -34,6 +34,7 @@ import networkx as nx
 from hcga.io import read_graphfile
 from hcga.Operations.operations import Operations
 import hcga.feature_analysis as hcga_analysis
+import hcga.plotting as hcga_plotting
 
 from tqdm import tqdm
 import time
@@ -322,331 +323,52 @@ class Graphs():
         X = self.X_norm
         y = self.y
 
+        feature_names = hcga_analysis.define_feature_set(self, data)
 
-        # defining the feature set
-        feature_names=[col for col in self.graph_feature_matrix.columns]            
-
-        if data != 'all':
-            matching_NL = [i for i, j in enumerate(feature_names) if 'NL' in j]
-            matching_NF = [i for i, j in enumerate(feature_names) if 'NF' in j]
-            matching = matching_NL + matching_NF
-            
-        if data=='topology':
-            X = np.delete(X,matching,axis=1)
-            feature_names = [i for j, i in enumerate(feature_names) if j not in matching]
-
-        elif data=='feats':
-            X = X[:,matching]
-            feature_names = [i for j, i in enumerate(feature_names) if j in matching]
-
-
-
-        testing_accuracy, top_feats = classification(X,y,ml_model)
+        testing_accuracy, top_feats = hcga_analysis.classification(X,y,ml_model)
         print("Mean test accuracy full set: --- {0:.3f} ---)".format(np.mean(testing_accuracy)))            
 
         # get names of top features
-        top_features_list = top_features(X,top_feats, feature_names, image_folder = image_folder, threshold = reduc_threshold, plot=plot)  
+        top_features_list, top_feat_indices = hcga_analysis.compute_top_features(X,top_feats,feature_names)
+
         self.top_feats = top_feats         
-        self.top_features_list=top_features_list       
+        self.top_features_list = top_features_list       
+
+        if plot:
+            hcga_plotting.plot_top_features(X,top_feats, feature_names, image_folder = image_folder, threshold = reduc_threshold)
+        #top_features(X,top_feats, feature_names, image_folder = image_folder, threshold = reduc_threshold, plot=plot)  
             
         # re running classification with reduced feature set
-        X_reduced, top_feat_indices = reduce_feature_set(X,top_feats, threshold = reduc_threshold)        
-        testing_accuracy_reduced_set, top_feats_reduced_set = classification(X_reduced,y,ml_model) 
+        X_reduced, top_feat_indices = hcga_analysis.reduce_feature_set(X,top_feats, threshold = reduc_threshold)        
+        testing_accuracy_reduced_set, top_feats_reduced_set = hcga_analysis.classification(X_reduced,y,ml_model) 
         print("Final mean test accuracy reduced feature set: --- {0:.3f} ---)".format(np.mean(testing_accuracy_reduced_set)))            
         
         # top and violin plot top features
         if plot:
-            self.top_features_importance_plot(X,top_feat_indices,feature_names,y, name='xgboost', image_folder = image_folder)
-            self.plot_violin_feature(X,y,top_feat_indices[0],feature_names, name='xgboost', image_folder = image_folder)
+            hcga_plotting.top_features_importance_plot(self, X,top_feat_indices,feature_names,y, name='xgboost', image_folder = image_folder)
+            hcga_plotting.plot_violin_feature(self,X,y,top_feat_indices[0],feature_names, name='xgboost', image_folder = image_folder)
 
 
         # univariate classification on top features
-        univariate_topfeat_acc = univariate_classification(X_reduced,y)
+        univariate_topfeat_acc = hcga_analysis.univariate_classification(X_reduced,y)
         feature_names_reduced = [feature_names[i] for i in top_feat_indices]
         top_feat_index = np.argsort(univariate_topfeat_acc)[::-1]       
 
         # top and violin plot top features from univariate
         if plot:
-            self.top_features_importance_plot(
+            hcga_plotting.top_features_importance_plot(self,
                 X_reduced,top_feat_index[0:2],
                 feature_names_reduced,y, name='univariate', 
                 image_folder = image_folder)
-            self.plot_violin_feature(
+            hcga_plotting.plot_violin_feature(self,
                 X_reduced,y,top_feat_index[0],feature_names_reduced, 
                 name='univariate', image_folder = image_folder)
 
         self.test_accuracy = testing_accuracy_reduced_set 
 
         return np.mean(testing_accuracy_reduced_set)
-    
-    
-    
-    def graph_regression(self,plot=True, data='all'):
-        """
-        Perform graph regression
-        """ 
-
-        from sklearn.model_selection import StratifiedKFold   
-        import xgboost
-        from sklearn.metrics import explained_variance_score
-
-               
-        hcga_analysis.normalise_feature_data(self)
-
-        X = self.X_norm
-        y = self.y
-        
-        feature_names=[col for col in self.graph_feature_matrix.columns]            
-
-        X = np.delete(X,np.where(np.isnan(y))[0],axis=0)
-        y = np.delete(y,np.where(np.isnan(y))[0],axis=0)
-        
-        # Let's try XGboost algorithm to see if we can get better results
-        xgb = xgboost.XGBRegressor(n_estimators=100, learning_rate=0.08,gamma=0,subsample=0.75,colsample_bytree=1, max_depth=7)
-                
-        bins = np.logspace(np.min(y), np.max(y), 10) 
-        y_binned = np.digitize(y, bins)
-        
-        skf = StratifiedKFold(n_splits = 10, shuffle = True)        
-        
-        top_feats = []        
-        explained_variance = []
-        for train_index, test_index in skf.split(X, y_binned):            
-            X_train, X_test = X[train_index], X[test_index]
-            y_train, y_test = y[train_index], y[test_index]
-        
-            xgb.fit(X_train,y_train)
-        
-            y_pred = xgb.predict(X_test)
-            
-            explained_variance.append(explained_variance_score(y_test,y_pred))
-
-            print("Fold explained variance: --- {0:.3f} ---)".format(explained_variance_score(y_test,y_pred)))            
-            top_feats.append(xgb.feature_importances_)
-
-        print("Final mean explained variance: --- {0:.3f} ---)".format(np.mean(explained_variance)))            
-        print("Final .std explained variance: --- {0:.3f} ---)".format(np.std(explained_variance)))           
-
-        top_features_list = top_features(X,top_feats,feature_names)            
-        self.top_features_list=top_features_list            
-
-            
-        if plot==True:
-            self.top_features_importance_plot(X,top_feats,y)
-    
-    def graph_classification_mlp(self,X = None, y = None , verbose=True):
-        """
-        Classify graphs with MLP algorithm
-        """
-
-        #from sklearn.model_selection import train_test_split
-        from sklearn.preprocessing import LabelBinarizer        
-        from sklearn.model_selection import StratifiedKFold          
-        import numpy as np
-        import tensorflow as tf
-        from sklearn.metrics import accuracy_score
-        
-        
-        hcga_analysis.normalise_feature_data(self)
-
-        if X is None:
-            X = self.X_norm
-            y = self.y        
-       
-        testing_accuracy = []
-        
-        
-        counts = np.bincount(y)
-        least_populated_class = np.argmin(counts)
-        if least_populated_class<10:
-            skf = StratifiedKFold(n_splits=len(y[y==least_populated_class]), random_state=10, shuffle=True)
-        else:
-            skf = StratifiedKFold(n_splits=10, random_state=10, shuffle=True)
-        
-        
-        for train_index, test_index in skf.split(X, y):
-            
-            X_train, X_test = X[train_index], X[test_index]
-            y_train, y_test = y[train_index], y[test_index]
-            #X_train, y_train, X_test, y_test = load_dataset(X,y)
-            
-            ## Changing labels to one-hot encoded vector
-            if len(np.unique(y_train)) > 2:
-                lb = LabelBinarizer()
-                y_train = lb.fit_transform(y_train)
-                y_test = lb.transform(y_test)
-            else:
-                y_train = np.array([y_train, 1-y_train]).T
-                y_test = np.array([y_test, 1-y_test]).T
-        
-            print('Train labels dimension:');print(y_train.shape)
-            print('Test labels dimension:');print(y_test.shape)            
-            
-
-            s = tf.Session()  # Create new session            
-            
-            ## Defining various initialization parameters for 600-256-128-# MLP model
-            num_classes = y_train.shape[1]
-            num_features = X_train.shape[1]
-            num_output = y_train.shape[1]
-            num_layers_0 = 256
-            num_layers_1 = 128#128
-            starter_learning_rate = 0.001
-            regularizer_rate = 0.1#0.1
-            
-            # Placeholders for the input data
-            input_X = tf.placeholder('float32',shape =(None,num_features),name="input_X")
-            input_y = tf.placeholder('float32',shape = (None,num_classes),name='input_Y')
-            
-            ## for dropout layer
-            keep_prob = tf.placeholder(tf.float32)
-            
-            
-            weights_0 = tf.Variable(tf.random_normal([num_features,num_layers_0], stddev=(1/tf.sqrt(float(num_features)))))
-            bias_0 = tf.Variable(tf.random_normal([num_layers_0]))
-            
-            weights_1 = tf.Variable(tf.random_normal([num_layers_0,num_layers_1], stddev=(1/tf.sqrt(float(num_layers_0)))))
-            bias_1 = tf.Variable(tf.random_normal([num_layers_1]))
-            
-            weights_2 = tf.Variable(tf.random_normal([num_layers_1,num_output], stddev=(1/tf.sqrt(float(num_layers_1)))))
-            bias_2 = tf.Variable(tf.random_normal([num_output]))
-            
-            hidden_output_0 = tf.nn.relu(tf.matmul(input_X,weights_0)+bias_0)
-            hidden_output_0_0 = tf.nn.dropout(hidden_output_0, keep_prob)
-            
-            hidden_output_1 = tf.nn.relu(tf.matmul(hidden_output_0_0,weights_1)+bias_1)
-            hidden_output_1_1 = tf.nn.dropout(hidden_output_1, keep_prob)
-            
-            predicted_y = tf.sigmoid(tf.matmul(hidden_output_1_1,weights_2) + bias_2)
-            
-            loss = tf.reduce_mean(tf.nn.softmax_cross_entropy_with_logits_v2(logits=predicted_y,labels=input_y)) + regularizer_rate*(tf.reduce_sum(tf.square(bias_0)) + tf.reduce_sum(tf.square(bias_1)))
-            
-            learning_rate = tf.train.exponential_decay(starter_learning_rate, 0, 5, 0.85, staircase=True)
-            
-            
-            ## Adam optimzer for finding the right weight
-            optimizer = tf.train.AdamOptimizer(learning_rate).minimize(loss,var_list=[weights_0,weights_1,weights_2,bias_0,bias_1,bias_2])
-            
-            ## Metrics definition
-            correct_prediction = tf.equal(tf.argmax(y_train,1), tf.argmax(predicted_y,1))
-            accuracy = tf.reduce_mean(tf.cast(correct_prediction, tf.float32))
-    
-            ## Training parameters
-            batch_size = 128
-            epochs=200
-            dropout_prob = 0.6
-            training_accuracy = []
-            training_loss = []
-            validation_accuracy = []
-            s.run(tf.global_variables_initializer())
-            for epoch in range(epochs):    
-                arr = np.arange(X_train.shape[0])
-                np.random.shuffle(arr)
-                for index in range(0,X_train.shape[0],batch_size):
-                    s.run(optimizer, {input_X: X_train[arr[index:index+batch_size]],
-                                      input_y: y_train[arr[index:index+batch_size]],
-                                    keep_prob:dropout_prob})
-                training_accuracy.append(s.run(accuracy, feed_dict= {input_X:X_train, 
-                                                                     input_y: y_train,keep_prob:1}))
-                training_loss.append(s.run(loss, {input_X: X_train, 
-                                                  input_y: y_train,keep_prob:1}))
-            
-                ## Evaluation of model
-                validation_accuracy.append(accuracy_score(y_test.argmax(1), 
-                                        s.run(predicted_y, {input_X: X_test,keep_prob:1}).argmax(1)))
-                if verbose:
-                    print("Epoch:{0}, Train loss: {1:.2f} Train acc: {2:.3f}, Val acc:{3:.3f}".format(epoch,
-                                                                                    training_loss[epoch],
-                                                                                    training_accuracy[epoch],
-                                                                                    validation_accuracy[epoch]))
-            
-            
-            #test_acc = accuracy_score(y_test.argmax(1),s.run(predicted_y, {input_X: X_test,keep_prob:1}).argmax(1))  
-            testing_accuracy.append(validation_accuracy[-50:])   
-            
-            #print("Test acc:{0:.3f}".format(test_acc))   
-            tf.reset_default_graph()
-                        
-        print("Final mean test accuracy: --- {0:.3f} ---)".format(np.mean(testing_accuracy)))            
-        
-        s.close()
-            
-        self.mlp_test_accuracy = testing_accuracy
-        
-        return np.mean(testing_accuracy)
         
 
-    def univariate_top_features(self):
-        """
-        Compute the univariate classification accuracies
-        """
-
-        hcga_analysis.normalise_feature_data(self)
-
-        X = self.X_norm
-        y = self.y
-        
-        classification_accs = univariate_classification(X,y)
-        
-        self.univariate_classification_accuracy = classification_accs
-        
-    def pca_features_plot(self,X,y,indices): 
-        """
-        Compute the PCA of the feature set and plot it
-        """
-
-        from sklearn.decomposition import PCA
-        import matplotlib.cm as cm  
-        pca = PCA(n_components=2)
-        
-        X1 = X[np.argsort(y),:]
-        y1 = y[np.argsort(y)]
-        
-        X_pca = pca.fit_transform(X1[:,indices]) 
-        cm = cm.get_cmap('RdYlBu') 
-        plt.scatter(X_pca[:,0],X_pca[:,1],cmap=cm,c=y1)                
-        #plt.ylim([-30,30])
-        #plt.xlim([-30,30])
-        plt.xlabel('PC1')        
-        plt.ylabel('PC2')
-        
-    def top_features_importance_plot(self,X,top_feat_indices,feature_names,y, name = 'xgboost', image_folder='images'):
-        """ 
-        Plot the top feature importances
-        """
-
-        import matplotlib.cm as cm  
-        import random
-        #mean_importance = np.mean(np.asarray(top_feats),0)                  
-        #top_feat_indices = np.argsort(mean_importance)[::-1]  
-        plt.figure()
-        cm = cm.get_cmap('RdYlBu')                  
-        sc = plt.scatter(X[:,top_feat_indices[0]],X[:,top_feat_indices[1]],cmap=cm,c=y)
-        plt.xlabel(feature_names[top_feat_indices[0]])        
-        plt.ylabel(feature_names[top_feat_indices[1]])
-        plt.colorbar(sc)
-        plt.savefig(image_folder + '/scatter_top2_feats_'+self.dataset+'_'+name+'.svg', bbox_inches = 'tight') 
-        
-    def plot_violin_feature(self,X,y,feature_id,feature_names, name = 'xgboost', image_folder='images'):   
-        """
-        Plot the violins of a feature
-        """
-
-        import random
-        feature_data = X[:,feature_id]
-        
-        data_split = []
-        for k in np.unique(y):
-            indices = np.argwhere(y==k)
-            data_split.append(feature_data[indices])        
-        
-        import seaborn as sns
-        plt.figure()
-        sns.set(style="whitegrid")
-        ax = sns.violinplot(data=data_split,palette="muted",width=1)
-        ax.set(xlabel='Class label', ylabel=feature_names[feature_id])
-        plt.savefig(image_folder + '/violin_plot_'+self.dataset+'_'+feature_names[feature_id]+'_'+name+'.svg', bbox_inches = 'tight') 
-        
     def save_feature_set(self,filename = 'Outputs/feature_set.pkl'):
         """
         Save the features in a pickle
@@ -694,60 +416,6 @@ class Graphs():
         self.graph_feature_matrix = feature_matrix
 
 
-def classification(X,y,ml_model, verbose=True):
-    """
-    Perform classification of a normalized feature data
-    """
-
-    from sklearn.model_selection import StratifiedKFold  
-    from sklearn.metrics import accuracy_score
-    
-    # reducing number of folds to half the least populated class
-    # e.g. if only 9 elements of class A then we only use int(9/2)=4 folds
-    counts = np.bincount(y)
-    n_splits = int(np.min(counts[counts>0])/2)
-
-    if n_splits < 2:
-        n_splits = 2
-        if verbose:
-            print('Small dataset, we only do ', n_splits, ' splits.')
-
-    elif n_splits > 10:
-        n_splits = 10
-    
-    skf = StratifiedKFold(n_splits=n_splits, random_state=10, shuffle=True)
-         
-    testing_accuracy = []
-    
-    if ml_model =='random_forest':
-        from sklearn.ensemble import RandomForestClassifier
-        model = RandomForestClassifier(n_estimators=100,max_depth=30)
-    elif ml_model == 'xgboost':
-        from xgboost import XGBClassifier
-        model = XGBClassifier(max_depth=4)
-        
-        
-    top_feats = []
-    
-    for train_index, test_index in skf.split(X, y):
-        
-        X_train, X_test = X[train_index], X[test_index]
-        y_train, y_test = y[train_index], y[test_index]
-        y_pred = model.fit(X_train,y_train).predict(X_test)
-        
-        acc = accuracy_score(y_test,y_pred)
-        
-        if verbose:
-            print("Fold test accuracy: --- {0:.3f} ---)".format(acc))            
-
-        testing_accuracy.append(acc)         
-        
-        top_feats.append(model.feature_importances_)
-
-    return testing_accuracy, top_feats
-    
-
-
 def calculate_features_single_graph(calc_speed, G):
     """
     Calculate the feature of a single graph, for parallel computations
@@ -759,112 +427,4 @@ def calculate_features_single_graph(calc_speed, G):
     return G_operations
 
 
-def univariate_classification(X,y):
-    """
-    Apply a univariate classification on each feature
-    """
-    
-    classification_acc = []
-    for i in range(X.shape[1]):
-        testing_accuracy, top_feats = classification(X[:,i].reshape(-1,1),y,'xgboost',verbose=False)
-        classification_acc.append(np.mean(testing_accuracy))
-        
-    return classification_acc
-    
 
-def top_features(X,top_feats,feature_names, image_folder, threshold = 0.9, plot=True):
-    """
-    Select and plot the dendogram, heatmap and importance distribution of top features
-    """
-
-    import pandas as pd
-    from scipy.cluster.hierarchy import dendrogram, linkage
-    import seaborn as sns
-    
-    mean_importance = np.mean(np.asarray(top_feats),0)                  
-    sorted_mean_importance = np.sort(mean_importance)[::-1]     
-
-    top_feat_indices = np.argsort(mean_importance)[::-1]            
-    top_features_list = []
-    for i in range(len(top_feat_indices)):                
-        top_features_list.append(feature_names[top_feat_indices[i]])    
-        
-    top_features_list=list(dict.fromkeys(top_features_list))    
-    top_features_list[:40]
-    df_top40 = pd.DataFrame(columns = top_features_list[:40], data=X[:,top_feat_indices[:40]])            
-    cor = np.abs(df_top40.corr())        
-    Z = linkage(cor, 'ward')   
-    
-    if plot:
-        plt.figure()
-        dn = dendrogram(Z)         
-        plt.savefig(image_folder+'/endogram_top40_features.svg', bbox_inches = 'tight') 
-
-        new_index = [int(i) for  i in dn['ivl']]
-        top_feats_names = [top_features_list[i] for i in new_index]             
-        df = df_top40[top_feats_names]
-        cor2 = np.abs(df.corr())            
-        plt.figure()
-        sns.heatmap(cor2, linewidth=0.5)
-        plt.savefig(image_folder+'/heatmap_top40_feature_dependencies.svg', bbox_inches = 'tight') 
-
-        # Taking only features till we have reached 90% importance
-        sum_importance = 0      
-        final_index = 0 
-        for i in range(len(sorted_mean_importance)):
-            sum_importance = sum_importance + sorted_mean_importance[i]
-            if sum_importance > threshold:
-                final_index = i
-                break
-        if final_index < 3: #take top 2 if no features are selected
-            final_index = 3
-
-
-        plt.figure()
-        plt.plot(np.sort(mean_importance)[::-1])
-
-        plt.xlabel('Features')
-        plt.ylabel('Feature Importance')
-        plt.xscale('log')
-        plt.yscale('symlog', nonposy='clip', linthreshy=0.001)    
-        plt.axvline(x=final_index,color='r')
-        plt.savefig(image_folder+'/feature_importance_distribution.svg', bbox_inches = 'tight') 
-    
-    #import pickle as pkl        
-    #pkl.dump(np.sort(mean_importance)[::-1], open('importance_data/'+image_folder+'.pkl','wb'), pkl.HIGHEST_PROTOCOL)
-
-    return top_features_list
-
-def reduce_feature_set(X,top_feats, threshold=0.9):
-    """
-    Reduce the feature set
-
-
-    Parameters
-    ---------
-    top_feats: list
-        List of features to keep
-        
-    """
-
-    mean_importance = np.mean(np.asarray(top_feats),0)   
-    sorted_mean_importance = np.sort(mean_importance)[::-1]     
-    
-    
-    # Taking only features till we have reached 90% importance
-    sum_importance = 0      
-    final_index = 0 
-    for i in range(len(sorted_mean_importance)):
-        sum_importance = sum_importance + sorted_mean_importance[i]
-        if sum_importance > threshold:
-            final_index = i
-            break
-    if final_index < 3: #take top 2 if no features are selected
-        final_index = 3
-
-
-    top_feat_indices = np.argsort(mean_importance)[::-1][:final_index]     
-        
-    X_reduced = X[:,top_feat_indices]
-    
-    return X_reduced, top_feat_indices
